Move setRoot to a dedicated RootAnimator
