+++ lib/android/app/src/main/java/com/reactnativenavigation/viewcontrollers/navigator/RootAnimator.kt
@@ -1,0 +1,29 @@
+package com.reactnativenavigation.viewcontrollers.navigator
+
+import android.animation.Animator
+import android.animation.AnimatorListenerAdapter
+import android.content.Context
+import android.view.View
+import com.reactnativenavigation.options.AnimationOptions
+import com.reactnativenavigation.viewcontrollers.common.BaseAnimator
+import com.reactnativenavigation.views.element.TransitionAnimatorCreator
+
+open class RootAnimator @JvmOverloads constructor(
+        context: Context,
+        private val transitionAnimatorCreator: TransitionAnimatorCreator = TransitionAnimatorCreator()
+) : BaseAnimator(context) {
+    open fun setRoot(root: View, setRoot: AnimationOptions, onAnimationEnd: Runnable) {
+        root.visibility = View.INVISIBLE
+        val set = setRoot.getAnimation(root)
+        set.addListener(object : AnimatorListenerAdapter() {
+            override fun onAnimationStart(animation: Animator) {
+                root.visibility = View.VISIBLE
+            }
+
+            override fun onAnimationEnd(animation: Animator) {
+                onAnimationEnd.run()
+            }
+        })
+        set.start()
+    }
+}
